requirements.txt
# Python依赖库

# HTTP请求库
requests>=2.28.0

# 数据处理库
pandas>=1.5.0

# SeekDB客户端（根据实际情况调整版本）
seekdb-client>=1.0.0

# 环境变量管理
python-dotenv>=0.20.0

# JSON处理
simplejson>=3.17.6

# 日志管理
python-json-logger>=2.0.7